+++ src/components/Dashboard/Dashboard.tsx
@@ -16,7 +16,7 @@ import './dashboard.scss';
 export default () => (
     <div className="dashboard-container">
         <div className="dashboard">
-            <Alert variant="info" label="Experimental release!">
+            <Alert variant="info" label="Experimental release! ">
                 This is an unsupported, experimental preview and it is subject
                 to major redesigns in the future.
             </Alert>