+++ src/components/EventChart/Cards/Device.tsx
@@ -28,13 +28,13 @@ export default () => {
         dataProfile,
         ltemTXReduction,
         nbiotTXReduction,
+        modemSystemPreference,
+        modemSupportLTEM,
+        modemSupportNBIoT,
+        modemSupportGNSS,
     } = useSelector(getDashboardState);
 
     const fields: DashboardCardFields = {
-        'Funcational Mode': {
-            value: parseFunctionalMode(functionalMode),
-            commands: ['AT+CFUN'] as const,
-        },
         IMEI: {
             value: IMEI ?? 'Unknown',
             commands: ['AT+CGSN'] as const,
@@ -55,7 +55,7 @@ export default () => {
             value: currentBand ?? 'Unknown',
             commands: ['AT%XCBAND'] as const,
         },
-        'AVAILABLE BANDS': {
+        'SUPPORTED BANDS': {
             value: availableBands
                 ? formatAvailableBands(availableBands)
                 : 'Unknown',
@@ -69,6 +69,40 @@ export default () => {
             value: manufacturer ?? 'Unknown',
             commands: ['AT+CGMI'] as const,
         },
+        VOLTAGE: {
+            value: 'NOT IMPLEMENTED',
+            commands: [],
+        },
+        'Preferred Bearer': {
+            value: parsePreferredBearer(modemSystemPreference),
+            commands: ['AT%XSYSTEMMODE'] as const,
+        },
+        'ENABLED BEARERS': {
+            value: parseSupportedValue([
+                modemSupportLTEM ?? false,
+                modemSupportNBIoT ?? false,
+                modemSupportGNSS ?? false,
+            ]),
+            commands: ['AT%XSYSTEMMODE'] as const,
+        },
+        'BATTERY WARNING': {
+            value: 'NOT IMPLEMENTED',
+            commands: [],
+        },
+        'FUNCTIONAL MODE': {
+            value: parseFunctionalMode(functionalMode),
+            commands: ['AT+CFUN'] as const,
+        },
+        GNSS: {
+            value: 'NOT IMPLEMENTED',
+            commands: [],
+        },
+        OVERHEATING: {
+            value: 'NOT IMPLEMENTED',
+            commands: [],
+        },
+
+        // Should be removed:
         'LTE-M TX Reduction': {
             value: formatMode(ltemTXReduction) ?? 'Unknown',
             commands: [] as const,
@@ -119,4 +153,29 @@ const formatMode = (mode?: Mode) => {
         return mode;
     }
     return mode.map(band => `${band.band}: ${band.reduction}`).join(', ');
+};
+
+const parseSupportedValue = (supported: [boolean, boolean, boolean]) => {
+    if (!supported.some(value => value === true)) return 'Nothing supported';
+    const [lteM, nbIot, gnss] = supported;
+    return `${lteM === true ? 'LTE-M ' : ''}${nbIot === true ? 'NB-IoT' : ''}${
+        gnss === true ? ' GNSS' : ''
+    }`;
+};
+
+const parsePreferredBearer = (preferred: number) => {
+    switch (preferred) {
+        case 0:
+            return 'No Preference';
+        case 1:
+            return 'LTE-M';
+        case 2:
+            return 'NB-IoT';
+        case 3:
+            return 'Network Selection (LTE-M)';
+        case 4:
+            return 'Network Selection (LTE-M)';
+        default:
+            return 'Unknown';
+    }
 };